Check Synapse version before binding email (#61)
Old Synapse versions (0.99.3 and earlier) did not support having emails bound by password providers, and as such cause errors with the current ldap3 password provider, which expects this functionality.

This change first checks the running Synapse version, and only binds emails if a version of Synapse > 0.99.3 is running.

The only issue now though is if someone is running off the current Synapse develop. That means they'll have the ability to bind emails, but since 0.99.4/1.0.0 hasn't been tagged yet, they're still marketing themselves as 0.99.3. So emails will not be bound. This isn't a game-breaking issue, but adding emails afterwards can be a pain. To this end, we should warn people that they should not run the version of matrix-synapse-ldap3 with this PR, and Synapse /develop before 0.99.4/1.0.0 is tagged, unless they don't care about emails not being bound to their account on signup.